// src/Interfaces/interface.ts
export interface Pokemon {
    abilities: Ability[];
    id: number;
    name: string;
    moves: Move[];
    types: Type[];
    species: Species;
    location_area_encounters: LocationAreaEncounters;
    sprites: Sprites;
    evolution_chain: EvolutionChain;
    chain: Chain;
}

export interface Ability {
    ability:   Species;
    is_hidden: boolean;
    slot:      number;
}

export interface Type {
    slot: number,
    type: Species
}

export interface Species {
    name: string;
    url:  string;
}

export interface Move {
    move: Species;
}

export interface Sprites {
    other: Other;
}

export interface Other {
    "official-artwork": OfficialArtwork;
}

export interface OfficialArtwork {
    front_default: string;
    front_shiny:   string;
}

export interface LocationAreaEncounters {
    url:string | undefined;
}

export interface LocationArea {
    name: string;
}


export interface EvolutionChain {
    url: string;
}

export interface Chain {
    evolves_to:        Chain[];
    is_baby:           boolean;
    species:           Species;
}

export interface PokemonSpeciesData {
    flavor_text_entries: {
      flavor_text: string;
      language: {
        name: string;
      };
    }[];
  }

// Trimmed it down to what I need

// Some where generated by https://quicktype.io

// I formatted other ones based on the quicktype generation